requirements.txt
fastapi
uvicorn
python-dotenv
redis
pymongo
google-generativeai
groq
langchain
streamlit_mic_recorder
nest_asyncio
langchain-community
langchain-huggingface
pypdf
transformers
torch
python-multipart 
tqdm              
sentence-transformers
langchain-mongodb
langchain-google-genai